requirement.txt
python-dotenv
django
Pillow
ipython
psycopg2-binary
djangorestframework
rest_framework_simplejwt
coverage


